feat: add interface changes at login and logout
